+++ tests/tc_curl_download.rb
@@ -32,6 +32,7 @@ class TestCurbCurlDownload < Test::Unit::TestCase
   end
 
   def test_download_url_to_file_via_io
+    omit('fork not available on this platform') if NO_FORK || WINDOWS
     dl_url = "http://127.0.0.1:9129/ext/curb_easy.c"
     dl_path = File.join(Dir::tmpdir, "dl_url_test.file")
     reader, writer = IO.pipe
@@ -58,7 +59,7 @@ class TestCurbCurlDownload < Test::Unit::TestCase
     assert File.exist?(dl_path)
     assert_equal File.read(File.join(File.dirname(__FILE__), '..','ext','curb_easy.c')), File.read(dl_path)
   ensure
-    File.unlink(dl_path) if File.exist?(dl_path)
+    File.unlink(dl_path) if dl_path && File.exist?(dl_path)
   end
 
   def test_download_bad_url_gives_404